update calculated variables
